requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15
mysqlclient>=2.2
Pillow>=9.5
django-filter>=24.2
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
boto3>=1.28
django-storages>=1.14
redis>=5.0
celery>=5.4
stripe>=10.5
razorpay>=1.4
python-dotenv>=1.0
requests>=2.31
pytest>=7.4
pytest-django>=4.7
pytest-cov>=4.1
coverage>=7.3
